Use ES modules
I was planning to be conservative about modules and go CJS modules, but
already saw that the first library I want to add uses ESM, and others I
plan to use also do that and there's no documentation for CJS, so I
don't even know if them are dual.

So ESM it is!